Write Dashboard file (#10)
Write Dashboard file:
Write a Dashboard file that has a Dashboard to highlight the number of objects and their relative Delivery Security Phase and Security Progress Phase.
It contains also another slide with highlights about risks, problems and news I want to highlight.

* Security Report Powerpoint template

I am adding the Powerpoint Template file from which we are going to write the report.

* Forcing UTF-8 encoding on files.

* Write parts of the Report on Powerpoint

I am writing only parts of the report based on a Powerpoint generated from the Report Powerpoint template.

* Update Powerpoint Template file

* Create Chart and write it on the Slide.

* Update Powerpoint template

* Write 2 charts on the slide with sample datas.

I am writing 2 charts on the slide.
I am adding legends to them and data labels to the columns.
The charts are written with sample datas.

* Refactor chart creation functions 

Refactor the chart data creation.
Refactor the chart data legend creation function
Refactor the chart data legend column

* Update Powerpoint Template files

Update the potx and pptx files that we are relaying on to write the report.

* Rewrite the Powerpoint writing according to the new template

* Remove pyc files

* Add new .gitignore

* Update Powerpoint Template

* Adding new variables about Security Phase

* Adding Security Phase Counter method

* Writing Security Phase Counter to Powerpoint file.

Writing Security Phase Counter to Powerpoint file using the template text placeholder.

* Forcing UTF-8 on "writereport" file.

* Writing Lorem Ipsum on the Highlight Slide

Writing some fake data on the Highlight (#2) Slide.

* Ignore all *.xlsx files

* Update Powerpoint Templates

* Remove Objects phase counter writing on PPT file

Remove Objects phase counter writing on the first slide of the PPT file as we have a histogram to present it now.

* Rewrite data for histograms and replace them

Rewrite data for histograms and replace them according to the new template.

* Create Security Report Class Structure

Create classes to describe a Security Report complete object.

* Create the Security Report object from the Objects Dict Array

* Write the Highlights slide from the Security Report Object.

* Add Security Report classes files

* Move the calls for the Security Report object from the new created file

Move the calls for the Security Report object from the new created file on the createdata module.
Use the SecurityModule  elements  to add the Highlights objects.

* Add IoT Object Info class file

I am adding a new class that describe an IoT object and all its information.

* Adding the objects information to IoTObjectInfo object

Adding the objects information to IoTObjectInfo object instead of Dictionnaries.

* Adding a module that has helper methods for the SecurityReport

It helps to populate a SecurityReport from an array of IoT Objects.
It has helper methods for all members of a SecurityReport object.
It can also create a SecurityReport object from an array of IoT objects.

* Adding some required counters as members of the SecurityReport object

* Creating the SecurityReport from the utils functions

* Writing the Powerpoint file from the SecurityReport

Writing the Powerpoint file from the SecurityReport instead of the IoT objects array as it was before.

* Doc: Adding some functions documentation

* Adding xml utils file to read and write xml files

* Using the Xml utils file to access the content of xml files

Using the Xml utils file to access the content of xml files on Data Creation

* WiP : Writing and Reading Highlights from an external xml file

* Adding an xml file that is a SecurityReport template.

* Revert "Adding an xml file that is a SecurityReport template."

This reverts commit 42658957b90adb30c870d44a569d4223402ae27b.

* Renaming the folder 'PowerpointTemplates' to 'Templates'

* Deleting the 'PowerpointReportTemplate' folder and content

* Deleting unnecessary xml file in createdata as it is in xml_utils

* Saving the SecurityReport object to a specific xml file.

* Saving Security Report from a SecurityReport template file.

* Moving some functions of xml_utils as private ones.

* Move Destination folder to the variables module.

* Use report history file to fulfill the PPT report file.

* Move the CURRENT_DIR to the variables file.

* Add data from last month on the charts

Add data from last month on the charts of the PPT Dashboard slide.
Datas are retrieved from the saved Security report file from last month.

* Move the Security Report creation to the Object itself. 

The SecurityReport object can populate itself its members values from a list of IoTObjects instead of relying on other modules.

* Move the SecurityReport saving to the object itself

The SecurityReport object can save its content to a (xml) file from the object itself instead of entirely reli-ying on the xml_utils module. SecurityReport is using the xml_utils file only to manipulate xml files (adding or modifying values, reading files, writing files).

* Adding new global variables

I am adding the 'Templates' folder name, the 'Report HIstory' folder name and the Current Highlights file name.

* Reading Risks from last month from report history and writing it.

* Reading highlights and adding it to SecurityReport

Reading highlights from CurrentHighlights file and adding it to Security Report object members.

* Remove unnecessary console prints, comments and refactor some comments

* Rename Phase dictionnary to the new one

As we are now relying on the  SecurityReport, the Delivery Security Process phases are now 'Evaluation', 'Qualification', 'Audits', 'Testing' and 'In Life' instead of the folder names.