Handle less than 5 objects in the riskiest chart in the ppt report.
On the Powerpoint slide that describes the riskiest objects, it was hardcoded that the 5 riskiest objects are going to be displayed. But if we have less than 5 objects it has no exception catched.
So I handled when we have less than 5 objects. We then display all the objects (5 or less).